+++ sickrage/core/nameparser/regexes.py
@@ -66,16 +66,15 @@ normal_regexes = [
      -(?P<release_group>[^ -]+([. _-]\[.*\])?))?)?)?$              # Group
      '''),
     ('newpct',
-     # American Horror Story HDTV x264[Cap.408_409]SPANISH AUDIO -NEWPCT
-     # American Horror Story [HDTV][Cap.408][Espanol Castellano]
-     # American Horror Story HDTV x264[Cap.408]SPANISH AUDIO –NEWPCT)
+     # American Horror Story - Temporada 4 HDTV x264[Cap.408_409]SPANISH AUDIO -NEWPCT
+     # American Horror Story - Temporada 4 [HDTV][Cap.408][Espanol Castellano]
+     # American Horror Story - Temporada 4 HDTV x264[Cap.408]SPANISH AUDIO –NEWPCT)
      r'''
-     ^(?P<series_name>.+?).-.                    # Show name: American Horror Story
+     (?P<series_name>.+?).-.+\d{1,2}[ ,.]       # Show name: American Horror Story
      (?P<extra_info>.+)\[Cap\.                   # Quality: HDTV x264, [HDTV], HDTV x264
      (?P<season_num>\d{1,2})                     # Season Number: 4
      (?P<ep_num>\d{2})                           # Episode Number: 08
-     ((_\d{1,2}(?P<extra_ep_num>\d{2}))|.*\]).*  # Episode Number2: 09
-     -(?P<release_group>[^ -]+([. _-]\[.*\])?)$  # Group
+     ((_\d{1,2}(?P<extra_ep_num>\d{2}))|.*\])     # Episode number2: 09
      '''),
     ('fov',
      # Show_Name.1x02.Source_Quality_Etc-Group